services/llm_generator/requirements.txt
fastapi
uvicorn
python-dotenv
requests
redis